CS: Rename file name(for Issue#7)
